# api/openapi.yaml
openapi: 3.1.0
info:
  title: Katalog API
  description: >
    REST API сервиса Katalog. Источник правды для контрактов:
    клиентские типы генерируются из этого файла (orval).
    Аутентификация: сессионная httpOnly cookie `katalog_session`.
    Тенант-изоляция: любой чужой или несуществующий ресурс — 404.
  version: 0.5.0
servers:
  - url: /
paths:
...
  /api/v1/shops:
    post:
      operationId: createShop
      summary: Создать магазин (slug валидируется, системные слова запрещены)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug: { type: string, minLength: 3, maxLength: 63 }
                name: { type: string, maxLength: 200 }
                description: { type: string }
                contacts: { type: object }
      responses:
        "201":
          description: Магазин создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      operationId: listShops
      summary: Магазины текущего пользователя
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/shops/{shopID}:
    parameters:
      - $ref: "#/components/parameters/ShopID"
    get:
      operationId: getShop
      summary: Магазин по id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateShop
      summary: Обновить магазин (частично)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 200 }
                description: { type: string }
                contacts: { type: object }
                settings: { type: object }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteShop
      summary: Удалить магазин
      responses:
        "204": { description: Удалено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/public/shops/{slug}:
    get:
      operationId: getPublicShop
      summary: Шапка магазина и сетка не скрытых альбомов
      parameters:
        - $ref: "#/components/parameters/ShopSlug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [shop, albums]
                properties:
                  shop: { $ref: "#/components/schemas/PublicShop" }
                  albums:
                    type: array
                    items: { $ref: "#/components/schemas/PublicAlbum" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  parameters:
    ShopSlug:
      name: slug
      in: path
      required: true
      schema: { type: string }
    ShopID:
      name: shopID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Нет или истекла сессия / неверные креды
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ресурс не найден или принадлежит другому пользователю
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Конфликт (занятый email/slug)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Превышен rate-limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Машинный код ошибки }
        message: { type: string }
...
    Shop:
      type: object
      required: [id, slug, name, description, contacts, settings, status, plan,
                 billing_state, paid_until, storage_used, storage_max, max_photos]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        contacts: { type: object }
        settings: { type: object }
        status: { type: string, enum: [active, suspended, deleted] }
        plan: { $ref: "#/components/schemas/PlanID" }
        billing_state: { $ref: "#/components/schemas/BillingState" }
        paid_until: { type: [string, "null"], format: date-time }
        storage_used: { type: integer }
        storage_max: { type: integer }
        max_photos: { type: integer }
    PlanID:
      type: string
      enum: [free, basic, pro]
    BillingState:
      type: string
      description: >
        Биллинговое состояние: ok — всё работает; grace — оплата просрочена,
        загрузка заблокирована, витрина работает; suspended — витрина скрыта,
        контент сохраняется.
      enum: [ok, grace, suspended]
...
    PublicShop:
      type: object
      description: Публичная витрина — без email, storage_used, plan, status
      required: [id, slug, name, description, contacts, msg_template]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        contacts:
          type: object
          description: "Каналы связи: telegram/whatsapp/vk/max -> handle"
        msg_template:
          type: string
          description: Шаблон сообщения покупателя, плейсхолдер {caption}
    PublicAlbum:
      type: object
      required: [id, parent_id, title, photo_count]
      properties:
        id: { type: string, format: uuid }
        parent_id: { type: [string, "null"], format: uuid }
        title: { type: string }
        photo_count: { type: integer }
        cover_urls: { $ref: "#/components/schemas/PhotoUrls" }
...
    PhotoUrls:
      type: object
      description: Деривативы WebP (300/800/1600), отдаются с CDN-пути /media
      properties:
        thumb: { type: string }
        medium: { type: string }
        large: { type: string }
...